+++ src/+wrappers/GrayscaleImageWrapper.m
@@ -72,7 +72,7 @@ classdef GrayscaleImageWrapper < wrappers.BaseImageWrapper
         function imageData = GetBrightening(obj, highpass, alpha)
             arguments
                 obj wrappers.GrayscaleImageWrapper
-                highpass double
+                highpass uint8
                 alpha(1, 1) double
             end % arguments
